Jwt auth implemented

+++ controllers/contactController.js
@@ -3,7 +3,7 @@ const Contact = require("../models/contactModel")
 
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts)
 })
 
@@ -25,7 +25,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
     res.status(201).json({message:'create contact', data: contact})
 });
@@ -35,6 +36,10 @@ const updateContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User does not have permission")
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id, 
@@ -50,6 +55,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User does not have permission")
     }
     await Contact.deleteOne({_id: req.params.id});
     res.status(200).json({message:'delete contact'})
